// src/types/vagonsr.types.ts
export interface IVagonResponse {
	ndoc: string
	id: string
	nvag: number
	inside_load_code: string
	kodtvag: string
	gpod: string
	vesvag: string
	ves_grotp: string
	ves_proves: string
	dat_ceh_otpr: string
	dat_st_otpr: string
	dat_st_nazn: string
	dat_ceh_nazn: string
	usr_ceh_otpr: string
	usr_st_otpr: string
	usr_st_nazn: string
	usr_ceh_nazn: string
	num_doc: number
}